recipe/requirements.txt
aniso8601==8.0.0
backpack==0.1
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
cleo==0.6.8
click==7.1.2
clikit==0.6.2
coverage==5.2
crashtest==0.3.0
cryptography==2.9.2
Faker==0.8.18
Flask==1.1.2
Flask-Login==0.5.0
flask-orator==0.2.0
Flask-RESTful==0.3.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.5.0
MarkupSafe==1.1.1
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
pyaml==16.12.2
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylev==1.3.0
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2019.3
PyYAML==5.3.1
simplejson==3.17.0
six==1.15.0
text-unidecode==1.2
tzlocal==1.5.1
Werkzeug==1.0.1
wrapt==1.12.1
